combined the spot filled and keyerror input validations
created an array containg all of the valid moves (playerBoard keys) used the if in statement to loop through the array. if it contains the players move and the key value is a hyphen, it changes the hyphen to either x or o and increments i. otherwise it gives the blanket invalid move statement. the continue causes the the loop to restart without checking for a winner or switching players. I am THRILLED!!!